Replaced the old schema with the new one.
